Randomize Songs
Used the Random util to randomize 4 songs in the music player. Only 4
songs exist in the player, does not expand if one adds more songs.

I.E if there were 10 files in the player, it would only randomize the
first 4 since that is the state in which the scope of this project was
received.